requirements.txt
numpy
requests
av
keyboard
PyYAML
pygetwindow